+++ plugins/modules/add_ldap_group.py
@@ -17,6 +17,11 @@ author:
 short_description: Add ldap group information.
 description:
    - Add ldap group information on Inspur server.
+deprecated:
+   removed_in: 3.0.0
+   why: Merge functions into the M(inspur.sm.ldap_group) module.
+   alternative: Use M(inspur.sm.ldap_group) instead.
+   removed_from_collection: inspur.sm
 options:
     name:
         description: