+++ plugins/modules/edit_pass_user.py
@@ -17,6 +17,11 @@ author:
 short_description: Change user password.
 description:
    - Change user password on Inspur server.
+deprecated:
+   removed_in: 3.0.0
+   why: Merge functions into the M(inspur.sm.user) module.
+   alternative: Use M(inspur.sm.user) instead.
+   removed_from_collection: inspur.sm
 options:
     uname:
         description: