+++ plugins/modules/edit_smtp_dest.py
@@ -10,30 +10,40 @@ __metaclass__ = type
 
 DOCUMENTATION = '''
 ---
-module: del_user
-version_added: "0.1.0"
+module: edit_smtp_dest
+version_added: "1.2.0"
 author:
     - WangBaoshan (@ISIB-group)
-short_description: Delete user.
+short_description: Set SMTP information.
 description:
-   - Delete user on Inspur server.
-deprecated:
-   removed_in: 3.0.0
-   why: Merge functions into the M(inspur.sm.user) module.
-   alternative: Use M(inspur.sm.user) instead.
-   removed_from_collection: inspur.sm
+   - Set SMTP dest information on Inspur server.
+   - Only the M6 models support this feature.
 options:
-    uname:
+    id:
         description:
-            - User name.
-        type: str
+            - Email destination id.
+        choices: [1, 2, 3, 4]
+        type: int
         required: true
+    status:
+        description:
+            - Email enable.
+        choices: ['enable', 'disable']
+        type: str
+    address:
+        description:
+            - Email address.
+        type: str
+    description:
+        description:
+            - Description information.
+        type: str
 extends_documentation_fragment:
     - inspur.sm.ism
 '''
 
 EXAMPLES = '''
-- name: Delete user test
+- name: Smtp  dest test
   hosts: ism
   connection: local
   gather_facts: no
@@ -45,9 +55,18 @@ EXAMPLES = '''
 
   tasks:
 
-  - name: "Delete user"
-    inspur.sm.del_user:
-      uname: "wbs"
+  - name: "Set smtp dest information"
+    inspur.sm.edit_smtp_dest:
+      id: 1
+      status: "disable"
+      provider: "{{ ism }}"
+
+  - name: "Set smtp dest information"
+    inspur.sm.edit_smtp_dest:
+      id: 1
+      status: "enable"
+      address: "100.2.2.2"
+      description": "test"
       provider: "{{ ism }}"
 '''
 
@@ -70,7 +89,7 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.inspur.sm.plugins.module_utils.ism import (ism_argument_spec, get_connection)
 
 
-class User(object):
+class SMTP(object):
     def __init__(self, argument_spec):
         self.spec = argument_spec
         self.module = None
@@ -84,7 +103,7 @@ class User(object):
             argument_spec=self.spec, supports_check_mode=False)
 
     def run_command(self):
-        self.module.params['subcommand'] = 'deluser'
+        self.module.params['subcommand'] = 'setsmtpdest'
         self.results = get_connection(self.module)
         if self.results['State'] == 'Success':
             self.results['changed'] = True
@@ -101,11 +120,14 @@ class User(object):
 
 def main():
     argument_spec = dict(
-        uname=dict(type='str', required=True),
+        id=dict(type='int', required=True, choices=[1, 2, 3, 4]),
+        status=dict(type='str', required=False, choices=['enable', 'disable']),
+        address=dict(type='str', required=False),
+        description=dict(type='str', required=False),
     )
     argument_spec.update(ism_argument_spec)
-    user_obj = User(argument_spec)
-    user_obj.work()
+    smtp_obj = SMTP(argument_spec)
+    smtp_obj.work()
 
 
 if __name__ == '__main__':