+++ plugins/modules/gpu_info.py
@@ -10,30 +10,21 @@ __metaclass__ = type
 
 DOCUMENTATION = '''
 ---
-module: del_user
-version_added: "0.1.0"
+module: gpu_info
+version_added: "1.2.0"
 author:
     - WangBaoshan (@ISIB-group)
-short_description: Delete user.
+short_description: Get GPU information.
 description:
-   - Delete user on Inspur server.
-deprecated:
-   removed_in: 3.0.0
-   why: Merge functions into the M(inspur.sm.user) module.
-   alternative: Use M(inspur.sm.user) instead.
-   removed_from_collection: inspur.sm
-options:
-    uname:
-        description:
-            - User name.
-        type: str
-        required: true
+   - Get GPU information on Inspur server.
+   - Only the M6 models support this feature.
+options: {}
 extends_documentation_fragment:
     - inspur.sm.ism
 '''
 
 EXAMPLES = '''
-- name: Delete user test
+- name: GPU test
   hosts: ism
   connection: local
   gather_facts: no
@@ -45,9 +36,8 @@ EXAMPLES = '''
 
   tasks:
 
-  - name: "Delete user"
-    inspur.sm.del_user:
-      uname: "wbs"
+  - name: "Get gpu information"
+    inspur.sm.gpu_info:
       provider: "{{ ism }}"
 '''
 
@@ -70,7 +60,7 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.inspur.sm.plugins.module_utils.ism import (ism_argument_spec, get_connection)
 
 
-class User(object):
+class CPU(object):
     def __init__(self, argument_spec):
         self.spec = argument_spec
         self.module = None
@@ -81,13 +71,11 @@ class User(object):
         """Init module object"""
 
         self.module = AnsibleModule(
-            argument_spec=self.spec, supports_check_mode=False)
+            argument_spec=self.spec, supports_check_mode=True)
 
     def run_command(self):
-        self.module.params['subcommand'] = 'deluser'
+        self.module.params['subcommand'] = 'getgpu'
         self.results = get_connection(self.module)
-        if self.results['State'] == 'Success':
-            self.results['changed'] = True
 
     def show_result(self):
         """Show result"""
@@ -100,12 +88,10 @@ class User(object):
 
 
 def main():
-    argument_spec = dict(
-        uname=dict(type='str', required=True),
-    )
+    argument_spec = dict()
     argument_spec.update(ism_argument_spec)
-    user_obj = User(argument_spec)
-    user_obj.work()
+    cpu_obj = CPU(argument_spec)
+    cpu_obj.work()
 
 
 if __name__ == '__main__':