+++ plugins/modules/user.py
@@ -63,7 +63,7 @@ EXAMPLES = '''
   tasks:
 
   - name: "Add user"
-    inspur.sm.add_user:
+    inspur.sm.user:
       state: "present"
       uname: "wbs"
       upass: "admin"
@@ -72,7 +72,7 @@ EXAMPLES = '''
       provider: "{{ ism }}"
 
   - name: "Set user"
-    inspur.sm.add_user:
+    inspur.sm.user:
       state: "present"
       uname: "wbs"
       upass: "12345678"